refactor: Extract shared POSIX PTY logic into TerminalHost.Posix base class
Deduplicates environment building, PATH construction, NVM detection,
and command resolution from Mac/Linux PTY services into PosixPtyServiceBase.
Renames methods for clarity and uses ushort internally for terminal dimensions.
