+++ src/TerminalHost.Linux/TerminalHost.Linux.csproj
@@ -8,6 +8,7 @@
   </PropertyGroup>
 <ItemGroup>
     <ProjectReference Include="..\TerminalHost.Core\TerminalHost.Core.csproj" />
+    <ProjectReference Include="..\TerminalHost.Posix\TerminalHost.Posix.csproj" />
     <ProjectReference Include="../../lib/PtySharp/PtySharp.Linux/PtySharp.Linux.csproj" />
   </ItemGroup>
 </Project>
